+++ src/Type/Collection/Interfaces/CollectionInterface.php
@@ -24,7 +24,7 @@ interface CollectionInterface extends \Countable, \Iterator, \ArrayAccess
      *
      * @param iterable $items
      * @param bool $useKey (use key when appending, false by default)
-     *
+     * @throws \Exception
      * @return CollectionInterface
      */
     public function appendMany(iterable $items, bool $useKey=false) : CollectionInterface;
@@ -59,6 +59,7 @@ interface CollectionInterface extends \Countable, \Iterator, \ArrayAccess
      * Syntax sugar for unset($collection[$key]); or $collection->offsetUnset($key);
      *
      * @param $key
+     * @throws \Exception
      * @return CollectionInterface
      */
     public function remove($key) : CollectionInterface;
@@ -66,6 +67,7 @@ interface CollectionInterface extends \Countable, \Iterator, \ArrayAccess
     /**
      * Remove last appended item
      *
+     * @throws \Exception
      * @return CollectionInterface
      */
     public function removeLast() : CollectionInterface;
@@ -147,4 +149,17 @@ interface CollectionInterface extends \Countable, \Iterator, \ArrayAccess
      */
     public function removeByValue($value, bool $strict = true) : int;
 
+    /**
+     * Prepends an element to the collection if the collection has numeric keys the
+     * keys will be reordered, for example [0 => 'test', 1=>'test2'] then "unshifting" a value 'test3' will result in:
+     * [0 => 'test3', 1=>'test', 2=>'test2']
+     *
+     * if the unshifted element has a string $key, then the key will be replaced, and the item will be first in the array
+     *
+     * @param mixed $item
+     * @param string|int $key
+     * @return CollectionInterface
+     * @throws \Exception
+     */
+    public function unshift($item, $key=null) : CollectionInterface;
 }